refactor(subagents): split bootstrap and runtime helpers
