+++ .changeset/split-subagents-bootstrap-and-command-runtime-helpers.md
@@ -1,0 +1,5 @@
+---
+default: patch
+---
+
+refactor the subagents extension by splitting bootstrap/config helpers, async runtime monitoring, and slash-command registration into focused modules.
